pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
